+++ R/module-sleep_spiral.R
@@ -32,6 +32,11 @@ sleep_spiral_module_server <- function(id, epochs, epochs_colnames) {
       if (nrow(epochs) == 0) {
         return(NULL)
       }
+      col <- epochs_colnames()
+      shiny::validate(
+        shiny::need(!is.null(col$timestamp), "'timestamp' column was not specified."),
+        shiny::need(!is.null(col$sleep_stage), "'sleep_stage' column was not specified.")
+      )
       plot_sleep_spiral(
         epochs = epochs,
         color_by = input$colorby,