+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Ambient Viewer Changelog
 
+## 0.0.3 (23/05/2025)
+
+### New
+
+- Summary outputs from GGIR (part 5) in csv format can be loaded as Sessions
+
+### Fixed
+
+- Setting column names with "Set Session Columns" and "Set Epoch Columns" work correctly
+- Loading a CSV file that doesn't fit the pre-set formats doesn't crash the app anymore
+
 ## 0.0.2.1 (19/05/2025)
 
 Patch to fix sleep clock display on the shiny server. No visible changes for users.